% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlrmd.R
\name{ragtag_html}
\alias{ragtag_html}
\title{Personal Rmd Temlate. It only converts to HTML right now.}
\usage{
ragtag_html(number_sections = FALSE, ...)
}
\arguments{
\item{number_sections}{Should sections be numbered?}

\item{...}{additional arguments sent to \@code{html_document}}
}
\description{
It essentially uses the base rmarkdown html_document function
but adds in additional CSS and and a footer template file.
It essentially uses the base rmarkdown html_document function,
but adds in additional CSS and and a footer template file.
I haven't added much to this footer template so it's more of a
skeleton for later.
}
